service/runner/requirements.txt
# Copyright 2024 Google LLC.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

functions-framework==3.*
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-logging==3.11.2
google-cloud-pubsub==2.23.0
google-cloud-storage==2.18.2
google-api-python-client==2.86.0
